requirements.txt

streamlit==1.39.0
python-dotenv==1.0.1
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.24
langchain-groq==0.1.3
langchain-huggingface==0.0.3
faiss-cpu==1.8.0.post1
huggingface-hub==0.24.6
transformers==4.44.2
sentence-transformers==3.0.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
reportlab==4.2.2
pypdf==4.2.0
